correct missed Bloggers.scala migration to new file tree
